refactor(component): change the fetch of the cards to props
