+++ src/components/ListCardComponent.vue
@@ -13,30 +13,18 @@
 </template>
 
 <script setup lang="ts">
-import { useMessage } from 'naive-ui'
-import { onMounted, ref } from 'vue'
+import { ref } from 'vue'
 
 import CardComponent from '@/components/CardComponent.vue'
-import { useApi } from '@/composables/useApi.ts'
 import type { Card } from '@/types'
 
-const api = useApi()
-const message = useMessage()
-
-const cards = ref<Card[]>([])
-const selectedCardIds = ref<Set<number>>(new Set())
-
-const fetchCards = async () => {
-  try {
-    cards.value = await api.getCards()
-  } catch (e) {
-    if (e instanceof Error) {
-      message.error(e.message)
-    } else {
-      message.error('Erreur inconnue')
-    }
-  }
+interface ListCardProps {
+  cards: Card[]
 }
+
+defineProps<ListCardProps>()
+
+const selectedCardIds = ref<Set<number>>(new Set())
 
 const emit =
   defineEmits<(e: 'update-selected-cards', ids: Set<number>) => void>()
@@ -54,8 +42,6 @@ const onSelectCard = (card: Card, selected: boolean) => {
 const isCardSelected = (cardId: number) => selectedCardIds.value.has(cardId)
 const isCardDisabled = (cardId: number) =>
   !selectedCardIds.value.has(cardId) && selectedCardIds.value.size === 10
-
-onMounted(fetchCards)
 </script>
 
 <style scoped></style>
